feat: add caching
As some scans may take hours, this adds caching so that requests return
more quickly.

This is not something prometheus suggests, however it seems that a
long running scrape even with a proper timeout sometimes won't show up
in the results once it's completed if it took an extended period of
time.

This will now always return a cached response.
If a request is received and the cache has expired, it will kick off
another scrape but still return the previous cache.