feat: add health metrics and endpoints (#21)
A few new metrics have been added to give more insight into the health
of the service.

- `masscan_collectors_total` reports how many collectors have been
  configured.
- `masscan_scrapes_total` reports the total number of scrapes executed
  for a particular collector, broken out by success and failed results.
- `masscan_scrapes_failed_current` reports the current number of
  consecutive failed scrapes.
  The value is reset upon a successful scrape.
- `masscan_scrape_next_start_time` reports the next time a scrape will
  occur.

In addition to the new metrics, `/livez` and `/readyz` endpoints were
added.
- `/livez` always returns a 200 status code.
- `/readyz` will report a 200 status code if there is at least one
  collector configured and none of the collectors have exceeded an
unhealthy number of scrapes.

The unhealthy number of failed scrapes may be configured by setting
`server.unhealthy_failed_scrapes` in the config.
If no value is provided, `5` is used by default.
This can be disabled by setting the value to `0`.